util.ts -> common.ts т.к. не проходил автотест - extra.feature   из компонента убрал Спинера постой tab с городами   в страницу с логином добавил случайный город   при старте приложения дожидаюсь только получения статуса авторизации   загрузку загрузку предложений перенес на главную страницу и спиннер добавил
