+++ src/components/spinner/spinner.tsx
@@ -14,16 +14,6 @@ function Spinner(): JSX.Element {
       </header>
 
       <main className="page__main page__main--index page__main--index-empty">
-        <h1 className="visually-hidden">Cities</h1>
-        <div className="tabs">
-          <section className="locations container">
-            <ul className="locations__list tabs__list" >
-              <li className="locations__item locations__item-link tabs__item">
-                <span>&nbsp;</span>
-              </li>
-            </ul>
-          </section>
-        </div>
         <div className="cities">
           <div className="cities__places-container cities__places-container--empty container">
             <section className="cities__no-places">
